requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
alembic>=1.14.0
passlib[bcrypt]>=1.7.4
sqladmin>=0.20.0
email-validator>=2.2.0
PyJWT>=2.8.0
openpyxl>=3.1.0
